% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_diff_max_min}
\alias{calc_diff_max_min}
\title{Difference between max and min value}
\usage{
calc_diff_max_min(v, na.rm = TRUE)
}
\arguments{
\item{v}{Vector with numbers}

\item{na.rm}{Should NA be removed}
}
\value{
Difference between max and min value in v vector
}
\description{
Difference between max and min value
}
