% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_rmse}
\alias{calc_rmse}
\title{Calculate root-mean-square error}
\usage{
calc_rmse(error)
}
\arguments{
\item{error}{Vector of errors}
}
\value{
Value of root-mean-square error
}
\description{
Calculate root-mean-square error
}
