% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_ssr}
\alias{calc_ssr}
\title{Calculate sum of the square of the residuals}
\usage{
calc_ssr(real_values, pred_values)
}
\arguments{
\item{real_values}{Real values}

\item{pred_values}{Predicted values}
}
\value{
sum of the square of the residuals
}
\description{
Calculate sum of the square of the residuals
}
