% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_variance_poly_reg}
\alias{calc_variance_poly_reg}
\title{Calculate variance for polynomial regression model}
\usage{
calc_variance_poly_reg(ssr, degree, num_data_points)
}
\arguments{
\item{ssr}{sum of the square of the residuals}

\item{degree}{Degree of polynomial}

\item{num_data_points}{Number of data points}
}
\value{
variance for polynomial regression model
}
\description{
Calculate variance for polynomial regression model
}
