% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_models.R
\name{consumption_predtemp_model}
\alias{consumption_predtemp_model}
\title{Get accuarcy of model based on linear regession for consumption and
temperature}
\usage{
consumption_predtemp_model(data, days = 1:7,
  decision_attr_prefix = "decision_attr_",
  consumption_prefix = "consumption_", num_previous_days = 7,
  time_col = "end_time", real_temp_col_prefix = "real_temp_",
  pred_col_prefix = "pred_temp_")
}
\arguments{
\item{data}{Dataset.}

\item{days}{Number of days for which prediction should be made.}

\item{decision_attr_prefix}{Prefix of columns with decision attribute.}

\item{consumption_prefix}{Prfix of column with consumption.}

\item{num_previous_days}{On how many past days linear regression should be
made?}

\item{time_col}{Name of column with time.}

\item{real_temp_col_prefix}{Prefix of columns with real temperatures.}

\item{pred_col_prefix}{Prefix of columns with predicted temperatures.}
}
\value{
RMSE of this model for each horizon.
}
\description{
Get accuarcy of model based on linear regession for consumption and
temperature
}
