% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_models.R
\name{consumption_time_model}
\alias{consumption_time_model}
\title{Get accuarcy of model based on linear regession for consumption and time}
\usage{
consumption_time_model(data, days = 1:7,
  decision_attr_prefix = "decision_attr_",
  consumption_prefix = "consumption_", num_previous_days = 7)
}
\arguments{
\item{data}{Dataset.}

\item{days}{Number of days for which prediction should be made.}

\item{decision_attr_prefix}{Prefix of columns with decision attribute.}

\item{consumption_prefix}{Prfix of column with consumption.}

\item{num_previous_days}{On how many past days linear regression should be
made?}
}
\value{
RMSE of this model for each horizon.
}
\description{
Get accuarcy of model based on linear regession for consumption and time
}
