% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{end_percentage_prefix}
\alias{end_percentage_prefix}
\title{Prefix for end percentage columns}
\usage{
end_percentage_prefix()
}
\value{
Prefix for end percentage columns
}
\description{
Prefix for end percentage columns
}
