% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{features_dataset}
\alias{features_dataset}
\title{Create dataset with features for gradient boosting model}
\usage{
features_dataset(data)
}
\arguments{
\item{data}{Dataset}
}
\value{
Features dataset
}
\description{
Create dataset with features for gradient boosting model
}
