% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental_model.R
\name{gbm_for_each_dataset}
\alias{gbm_for_each_dataset}
\title{Use gradient boosting model for many datasets.}
\usage{
gbm_for_each_dataset(dataset_path = "./data/datasets/",
  feature_path = "./data/features/", pattern_datasets = "*",
  pattern_features = "*", read_file_function = readr::read_csv,
  ntrees = 200, min_num_days = 50, ...)
}
\arguments{
\item{dataset_path}{Directory with datasets.}

\item{feature_path}{Directory with features from datasets.}

\item{pattern_datasets}{Regular expression. Only file names which match the
regular expression will be used from directory \code{dataset_path}.}

\item{pattern_features}{Regular expression. Only file names which match the
regular expression will be used from directory \code{feature_path}.}

\item{read_file_function}{Function used for reading files from
\code{dataset_path} and \code{pattern_features}.}

\item{ntrees}{Number of trees used in models.}

\item{min_num_days}{Minimum number of rows based on which model can be
created.}

\item{...}{Additional parameters for function \code{read_file_function}.}
}
\value{
List with results of evaluation for every dataset.
}
\description{
Use gradient boosting model for many datasets.
}
