% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_gradient_boosting_report}
\alias{generate_gradient_boosting_report}
\title{Generate a report with an evaluation of gradient boosting models for
prediction of LPG usage}
\usage{

  generate_gradient_boosting_report(data_path = system.file("extdata/datasets/",
  package = "lpg.short.pred"),
  features_path = system.file("extdata/features/", package =
  "lpg.short.pred"), ntrees = 100, report_html_name = "gbm_report",
  report_html_directory = ".")
}
\arguments{
\item{data_path}{Character. An absolute path to a directory with datasets
containing basic information about consumption and temperature. See README
in directory \code{inst/extdata} of package \code{lpg.short.pred} for more
information about a structure of these files. The default value of this
parameter is set to a directory with datasets shared with this package.}

\item{features_path}{Character. An absolute path to a directory with feature
datasets. See README in directory \code{inst/extdata} of package
\code{lpg.short.pred} for more information about a structure of these
files. The default value of this parameter is set to a directory with
feature datasets shared with this package.}

\item{ntrees}{Number of trees used in models.}

\item{report_html_name}{Character. Name of a generated HTML file with a
report. Default value is \emph{gbm_report}.}

\item{report_html_directory}{Character. A directory where generated HTML
report should be saved.}
}
\value{
A path to a generated HTML report.
}
\description{
Generate a report with an evaluation of gradient boosting models for
prediction of LPG usage
}
\examples{
generate_gradient_boosting_report()
}
