% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_naive_models_report}
\alias{generate_naive_models_report}
\title{Generate a report with an evaluation of naive models for prediction of LPG
usage}
\usage{

  generate_naive_models_report(data_path = system.file("extdata/datasets/",
  package = "lpg.short.pred"), report_html_name = "naive_models",
  report_html_directory = ".")
}
\arguments{
\item{data_path}{Character. An absolute path to a directory with datasets
containing basic information about consumption and temperature. See README
in directory \code{inst/extdata} of package \code{lpg.short.pred} for more
information about a structure of these files. The default value of this
parameter is set to a directory with datasets shared with this package.}

\item{report_html_name}{Character. Name of a generated HTML file with a
report. Default value is \emph{naive_models}.}

\item{report_html_directory}{Character. A directory where generated HTML
report should be saved.}
}
\value{
A path to a generated HTML report.
}
\description{
This function generates an HTML report with an evaluation of trivial models
used for prediction of LPG consumption.
}
\examples{
generate_naive_models_report()
}
