% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_models.R
\name{get_pred_mean_temp}
\alias{get_pred_mean_temp}
\title{Get mean values of predicted temperatures in future days}
\usage{
get_pred_mean_temp(index, data, pred_col_prefix = "pred_temp_",
  num_days = 7, time_col = "end_time")
}
\arguments{
\item{index}{Index of row for 'data'.}

\item{data}{Dataset}

\item{pred_col_prefix}{Prefix of columns with predicted temperatures.}

\item{num_days}{Number of future days for which predicted mean temperature
for each day should be calculated.}

\item{time_col}{Name of column with time of a record.}
}
\value{
Vector of predicted mean temperatures for each of future days.
}
\description{
Get mean values of predicted temperatures in future days
}
