% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_models.R
\name{get_previous_temp_for_days}
\alias{get_previous_temp_for_days}
\title{Get mean values of temperatures in previous days}
\usage{
get_previous_temp_for_days(index, data,
  real_temp_col_prefix = "real_temp_", num_days = 7,
  time_col = "end_time")
}
\arguments{
\item{index}{Index of row for 'data'.}

\item{data}{Dataset}

\item{real_temp_col_prefix}{Prefix of columns with real temperatures.}

\item{num_days}{Number of previous days for which mean temperatures for each
day should be calculated.}

\item{time_col}{Name of column with time of a record.}
}
\value{
Vector of mean temperatures for each of previous days.
}
\description{
For each day mean temperature is calculated based on hourly values.
}
