% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental_model.R
\name{gradient_boosting_model}
\alias{gradient_boosting_model}
\title{Create gradient boosting model}
\usage{
gradient_boosting_model(train_features, decision, ntrees)
}
\arguments{
\item{train_features}{Dataset with features based on which model will be
created.}

\item{decision}{Vector with decisions (what should be predicted).}

\item{ntrees}{Number of trees used in gradient boosting model.}
}
\value{
Gradient boosting model
}
\description{
Create gradient boosting model
}
