% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_models.R
\name{naive_models_dataset}
\alias{naive_models_dataset}
\title{Get models accuarcy for one dataset}
\usage{
naive_models_dataset(data, days = 1:7, time_col = "end_time")
}
\arguments{
\item{data}{Dataset}

\item{days}{Number of days for which prediction should be made.}

\item{time_col}{Name of column with time.}
}
\value{
RMSE of all models for this dataset.
}
\description{
Get models accuarcy for one dataset
}
