% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{predicted_temp_prefix}
\alias{predicted_temp_prefix}
\title{Prefix for columns with predicted temperatures}
\usage{
predicted_temp_prefix()
}
\value{
Prefix for columns with predicted temperatures
}
\description{
Prefix for columns with predicted temperatures
}
