% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{previous_consumptions}
\alias{previous_consumptions}
\title{Get values of previous consumptions for defined number of days.}
\usage{
previous_consumptions(data, consumption_prefix = "consumption_",
  num_days = 7)
}
\arguments{
\item{data}{Dataset.}

\item{consumption_prefix}{Prefix of columns with consumption.}

\item{num_days}{Number of days for which consumption should be returned.}
}
\value{
Consumption of previous days.
}
\description{
Get values of previous consumptions for defined number of days.
}
