% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{real_temp_prefix}
\alias{real_temp_prefix}
\title{Prefix for columns with real temperatures}
\usage{
real_temp_prefix()
}
\value{
Prefix for columns with real temperatures
}
\description{
Prefix for columns with real temperatures
}
