% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datasets.R
\name{save_features_datasets}
\alias{save_features_datasets}
\title{Create datasets with features from files with data}
\usage{
save_features_datasets(dataset_path = "./data/datasets",
  features_path = "./data/features/",
  read_file_function = readr::read_csv, pattern_datasets = "*")
}
\arguments{
\item{dataset_path}{Directory where files wit data are stores.}

\item{features_path}{Directory where new files with features should be
stored.}

\item{read_file_function}{Function used for reading files from
\code{dataset_path}.}

\item{pattern_datasets}{Regular expression. Only file names which match the
regular expression will be used from directory \code{dataset_path}.}
}
\value{
Nothing
}
\description{
Create datasets with features from files with data
}
