% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental_model.R
\name{use_gbm_for_dataset}
\alias{use_gbm_for_dataset}
\title{Evaluate incremental gradient boosting models on a dataset.}
\usage{
use_gbm_for_dataset(dataset, features = NULL,
  min_examples_for_model = 50, ntrees = 200, test_ntrees = 200)
}
\arguments{
\item{dataset}{Original ataset}

\item{features}{Dataset with features}

\item{min_examples_for_model}{Minimum examples that should be used in
creation of a first model.}

\item{ntrees}{Number of trees used in training.}

\item{test_ntrees}{On which number of trees evaluation should be made?}
}
\value{
Dataframe. Every column is a different horizon. Every row is next
  iteration.
}
\description{
For each horizon (1 to 7 days) models are created. For each horizon models
are created in incremental way. For example: first model is created on 2
first months and test on third month, second model is trained on 3 months and
evalueated on 4th etc.
}
